requirements.txt
pytest-cov
python-coveralls
codecov
coverage
cython
RPi.GPIO
smbus2
bokeh
ujson
cbor
numpy
